Separate lexer from parser
