+++ src/support/CMakeLists.txt
@@ -2,6 +2,7 @@ target_sources(morpho
     PRIVATE
         common.c     common.h
         extensions.c extensions.h
+        lex.c        lex.h
         memory.c     memory.h
         parse.c      parse.h
         random.c     random.h
